type:refactor add notification_controller and rename alert_notification to setup_notification
